Agrego test faltantes: calcularDistancia(),devuelveLocalidadDeUnMunicipio(),devuelveMunicipioDeUnaProvicia()

+++ src/main/java/dds/grupo3/clases/services/geodds/entities/Distancia.java
@@ -1,6 +1,6 @@
 package dds.grupo3.clases.services.geodds.entities;
 
 public class Distancia {
-  public long valor;
+  public Double valor;
   public String unidad;
 }
